Replace source with the v0.1.0 production line plus Cloudflare PATCH fix
The previous contents of this branch were an abandoned line that did not match
what runs in production. This replaces them with the v0.1.0 source (tag
e2252fd), which production actually runs, plus two fixes:

- Cloudflare record updates now use PATCH instead of PUT. PUT is an overwrite,
  so Cloudflare resets any field omitted from the body to its default. That
  silently set proxied=false and disabled the proxy on proxied records.
- A failed update now logs Cloudflare's response body, so a 404 shows whether
  the zone or the record was not found instead of just reporting the status.

Adds docs/BUILDING.md covering the single-file publish, how to tell this line
apart from the config-subfolder line, and deployment and verification steps.

+++ src/DdnsUpdate.Core/Models/DdnsSettings.cs
@@ -2,8 +2,8 @@
 // <copyright file="DdnsSettings.cs" company="PaulTechGuy">
 // Copyright (c) Paul Carver. All rights reserved.
 // </copyright>
-// Use of this source code is governed by Apache License 2.0 that can
-// be found at https://www.apache.org/licenses/LICENSE-2.0.
+// Use of this source code is governed by an MIT-style license that can
+// be found in the LICENSE file or at https://opensource.org/licenses/MIT.
 // -------------------------------------------------------------------------
 
 namespace DdnsUpdate.Core.Models;
@@ -12,7 +12,7 @@ public class DdnsSettings
 {
    public const string ConfigurationName = "DdnsSettings";
 
-   public decimal AfterAllDdnsUpdatePauseMinutes { get; set; }
+   public int AfterAllDdnsUpdatePauseMinutes { get; set; }
 
    public bool AlwaysUpdateDdnsEvenIfUnchanged { get; set; } = true;
